Split up functions in plot.py into submodules in a plot folder
