+++ osmo_jupyter/plot/constants.py
@@ -1,0 +1,16 @@
+COLOR_CHANNELS = ['r', 'g', 'b']
+
+# Axis # to use for each color when colors are being plotted on separate axes
+AXES_BY_COLOR = {
+    'r': 2,
+    'g': 3,
+    'b': 4,
+}
+ANNOTATION_AXIS = 1
+SHARED_COLOR_AXIS = 2
+
+COLORS_BY_LETTER = {
+    'r': 'red',
+    'g': 'green',
+    'b': 'blue'
+}
